student management stystem
The goal of system design is to transform the information collected about the
project into the blueprint structure which will serve as a base while constructing
the system. It is considered to be an unwieldy process as most of the errors are
introduced in this phase.
However, if an error gets unnoticed in the later process it may become difficult to
track them down. In our system, we are developing a system that helps in student
management by the institution. It will minimize the problems faced by both
students and the institutes for the duration.